merge: ref/jpa-entity into backend

See merge request s09-final/S09P31C103!55

+++ back-end/src/main/java/com/ssafy/eoullim/repository/FollowRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface FollowRepository extends JpaRepository<FollowEntity, Integer> {
+public interface FollowRepository extends JpaRepository<FollowEntity, Long> {
 
     Optional<FollowEntity> findByChildAndFollowingChild(ChildEntity child, ChildEntity followingChild);
